api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
requests